Removed proxy listener and moved HTTPMutator
Now the HTTPMutator that handles the "Send to" actions is completely inside the sendto.py file
