+++ python/inql/burp_ext/contextual.py
@@ -1,25 +1,10 @@
-from argparse import Action
-import json
-import re
-
-try:
-    import urllib.request as urllib_request # for Python 3
-    from urllib.parse import urlencode
-except ImportError:
-    import urllib2 as urllib_request # for Python 2 and Jython
-    from urllib import urlencode
-
 import logging
 from java.awt.event import ActionListener
 from javax.swing import JMenuItem
 
-from org.python.core.util import StringUtil
+from burp import IContextMenuFactory
 
-from burp import IProxyListener, IContextMenuFactory
-
-from inql.actions.sendto import HTTPMutator
-from inql.utils import is_query, override_headers, string_join, override_uri, clean_dict, multipart, random_string, \
-    querify, json_encode
+from inql.utils import is_query
 
 class SendMenuItem(IContextMenuFactory):
     """Handles additional entries in context (right-click) menu in the "Send To ..." style.
@@ -134,175 +119,3 @@ class OmniMenuItem(IContextMenuFactory):
         except Exception as ex:
             return None
         return mymenu
-
-
-class BurpHTTPMutator(HTTPMutator, IProxyListener):
-    def __init__(self, callbacks=None, helpers=None, overrideheaders=None, requests=None, stub_responses=None, attacker_receiver=None):
-        super(BurpHTTPMutator, self).__init__(overrideheaders=overrideheaders, requests=requests, stub_responses=stub_responses)
-
-        if helpers and callbacks:
-            self._helpers = helpers
-            self._callbacks = callbacks
-            self._callbacks.registerProxyListener(self)
-            # for r in self._callbacks.getProxyHistory():
-            #     self._process_request(self._helpers.analyzeRequest(r), r.getRequest())
-
-    def processProxyMessage(self, messageIsRequest, message):
-        """
-        Implements IProxyListener method
-
-        :param messageIsRequest: True if BURP Message is a request
-        :param message: message content
-        :return: None
-        """
-        return None
-
-    def send_to_attacker(self, host, payload, action):
-        logging.debug("send_to_attacker(%s, %s, %s" % (host, payload, action))
-        req = self._requests[host]
-        if req and self._callbacks and self._helpers:
-            body = req['body']
-            info = self._helpers.analyzeRequest(body)
-
-            nobody = body[:info.getBodyOffset()]
-            rstripoffset = info.getBodyOffset()-len(nobody.rstrip())
-            headers = body[:info.getBodyOffset()-rstripoffset]
-           
-            try:
-                self._overrideheaders[host]
-            except KeyError:
-                self._overrideheaders[host] = []
-
-            headers = override_headers(headers, self._overrideheaders[host])
-            repeater_body = StringUtil.toBytes(string_join(
-                headers,
-                body[info.getBodyOffset()-rstripoffset:info.getBodyOffset()],
-                payload))
-
-            url = "%s://%s" % (req['scheme'], req['host'])
-            if req['port'] != None:
-                url = url + ":" + str(req['port'])
-            action(url, repeater_body, inql=True)
-
-    def send_to_repeater(self, host, payload):
-        # get the request
-        req = self._requests[host]
-        if req and self._callbacks and self._helpers:
-            body = req['body']
-            info = self._helpers.analyzeRequest(body)
-
-            nobody = body[:info.getBodyOffset()]
-            rstripoffset = info.getBodyOffset()-len(nobody.rstrip())
-            headers = body[:info.getBodyOffset()-rstripoffset]
-           
-            try:
-                self._overrideheaders[host]
-            except KeyError:
-                self._overrideheaders[host] = []
-
-            # override/add the custom headers to the default ones
-            headers = override_headers(headers, self._overrideheaders[host])
-            repeater_body = StringUtil.toBytes(string_join(
-                headers,
-                body[info.getBodyOffset()-rstripoffset:info.getBodyOffset()],
-                payload))
-           
-            self._callbacks.sendToRepeater(req['host'], int(req['port']),
-                                           req['scheme'] == 'https', repeater_body,
-                                          'GraphQL #%s' % self._index)
-            self._index += 1
-
-    def send_to_repeater_get_query(self, host, payload):
-        req = self._requests[host]
-        if req and self._callbacks and self._helpers:
-            body = req['body']
-            info = self._helpers.analyzeRequest(body)
-
-            nobody = body[:info.getBodyOffset()]
-            rstripoffset = info.getBodyOffset()-len(nobody.rstrip())
-            headers = body[:info.getBodyOffset()-rstripoffset]
-           
-            try:
-                self._overrideheaders[host]
-            except KeyError:
-                self._overrideheaders[host] = []
-
-            headers = override_headers(headers, self._overrideheaders[host])
-            # remove Content-Type on GET requests
-            headers = re.sub(r'(?m)^Content-Type:.*\n?', '', headers)
-            content = json.loads(payload)
-            if isinstance(content, list):
-                content = content[0]
-            headers = override_uri(headers, method="GET", query=urlencode(querify(clean_dict(content))))
-
-            repeater_body = StringUtil.toBytes(string_join(
-                headers,
-                body[info.getBodyOffset()-rstripoffset:info.getBodyOffset()]))
-
-            self._callbacks.sendToRepeater(req['host'], int(req['port']),
-                                           req['scheme'] == 'https', repeater_body,
-                                          'GraphQL - GET query #%s' % self._index)
-            self._index += 1
-
-    def send_to_repeater_post_urlencoded_body(self, host, payload):
-        req = self._requests[host]
-        if req and self._callbacks and self._helpers:
-            body = req['body']
-            info = self._helpers.analyzeRequest(body)
-
-            nobody = body[:info.getBodyOffset()]
-            rstripoffset = info.getBodyOffset()-len(nobody.rstrip())
-            headers = body[:info.getBodyOffset()-rstripoffset]
-           
-            try:
-                self._overrideheaders[host]
-            except KeyError:
-                self._overrideheaders[host] = []
-
-            headers = override_headers(headers, self._overrideheaders[host])
-            headers = override_headers(headers, [("Content-Type", "application/x-www-form-urlencoded")])
-            headers = override_uri(headers, method="POST")
-            content = json.loads(payload)
-            if isinstance(content, list):
-                content = content[0]
-            repeater_body = StringUtil.toBytes(string_join(
-                headers,
-                body[info.getBodyOffset()-rstripoffset:info.getBodyOffset()],
-                urlencode(querify(clean_dict(content)))))
-
-            self._callbacks.sendToRepeater(req['host'], int(req['port']),
-                                           req['scheme'] == 'https', repeater_body,
-                                          'GraphQL - POST urlencoded #%s' % self._index)
-            self._index += 1
-
-    def send_to_repeater_post_form_data_body(self, host, payload):
-        req = self._requests[host]
-        if req and self._callbacks and self._helpers:
-            body = req['body']
-            info = self._helpers.analyzeRequest(body)
-
-            nobody = body[:info.getBodyOffset()]
-            rstripoffset = info.getBodyOffset()-len(nobody.rstrip())
-            headers = body[:info.getBodyOffset()-rstripoffset]
-           
-            try:
-                self._overrideheaders[host]
-            except KeyError:
-                self._overrideheaders[host] = []
-
-            headers = override_headers(headers, self._overrideheaders[host])
-            boundary = "---------------------------%s" % random_string()
-            headers = override_headers(headers, [("Content-Type", "multipart/form-data, boundary=%s" % boundary)])
-            headers = override_uri(headers, method="POST")
-            content = json.loads(payload)
-            if isinstance(content, list):
-                content = content[0]
-            repeater_body = StringUtil.toBytes(string_join(
-                headers,
-                body[info.getBodyOffset()-rstripoffset:info.getBodyOffset()],
-                multipart(data=querify(clean_dict(content)), boundary=boundary)))
-
-            self._callbacks.sendToRepeater(req['host'], int(req['port']),
-                                           req['scheme'] == 'https', repeater_body,
-                                          'GraphQL - POST form-data #%s' % self._index)
-            self._index += 1
